requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.2
requests==2.32.3